requirements.txt
pdf2image==1.16.0
pytesseract==0.3.8
Pillow==8.4.0
Flask==2.1.2
Werkzeug==2.2.2